requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
emoji==2.10.1
greenlet==3.0.3
idna==3.6
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
PySocks==1.7.1
